[MODRR-17]-Added foreign key validation
