+++ src/main/java/org/folio/readingroom/exception/PatronPermissionException.java
@@ -1,0 +1,8 @@
+package org.folio.readingroom.exception;
+
+public class PatronPermissionException extends RuntimeException {
+
+  public PatronPermissionException(String errorMsg) {
+    super(errorMsg);
+  }
+}
