--- src/main/java/org/folio/readingroom/exception/PatronPermissionsException.java
@@ -1,8 +1,0 @@
-package org.folio.readingroom.exception;
-
-public class PatronPermissionsException extends RuntimeException {
-
-  public PatronPermissionsException(String errorMsg) {
-    super(errorMsg);
-  }
-}
