+++ src/main/java/org/folio/readingroom/service/PatronPermissionService.java
@@ -4,7 +4,7 @@ import java.util.List;
 import java.util.UUID;
 import org.folio.readingroom.domain.dto.PatronPermission;
 
-public interface PatronPermissionsService {
+public interface PatronPermissionService {
   List<PatronPermission> updatePatronPermissionsByUserId(UUID patronId,
     List<PatronPermission> patronPermissions);
